:recycle: refactor layer
