+++ src/main/java/potato/onetake/infrastructure/config/SecurityConfig.java
@@ -1,4 +1,4 @@
-package potato.onetake.global.config;
+package potato.onetake.infrastructure.config;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
